# openapi/cms-api.yaml
openapi: 3.1.0
info:
  title: CMS Collection Management System API
  version: 1.0.0
  description: |
    Multi-tenant Collection Management System API.
    
    ## Authentication
    All endpoints (except /health and /ready) require authentication via Bearer token.
    
    ## Multi-Tenancy
    All requests must include tenant context. In development, use the `x-tenant-id` header.
    In production, tenant is derived from the JWT token.
    
    ## Request Tracing
    Include `x-request-id` and `x-correlation-id` headers for distributed tracing.
  contact:
    name: CMS Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.cms.example.com
    description: Production server
...
paths:
...
  /api/v1/mdm/sources:
    get:
      operationId: listSourceSystems
      summary: List source systems
      tags:
        - MDM - Sources
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, deprecated]
      responses:
        '200':
          description: List of source systems
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createSourceSystem
      summary: Create source system
      tags:
        - MDM - Sources
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSourceSystem'
      responses:
        '201':
          description: Source system created

  # ========== EPIC 5: MDM - Reference Data ==========
...
  /api/v1/mdm/parties/{party_id}/sources:
    get:
      operationId: getPartySources
      summary: Get source mappings for a party
      tags:
        - MDM - Parties
      parameters:
        - name: party_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of source system mappings

  # ========== EPIC 5: MDM - Contracts ==========
  /api/v1/mdm/contracts:
    get:
      operationId: listContracts
      summary: Search and list contract golden records
      tags:
        - MDM - Contracts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: party_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of contracts
  /api/v1/mdm/contracts/{contract_id}:
    get:
      operationId: getContract
      summary: Get contract with sources and DQ issues
      tags:
        - MDM - Contracts
      parameters:
        - name: contract_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Contract details with related data

  # ========== EPIC 5: MDM - Users ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with tenant and user claims
...
  schemas:
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        data:
          description: Response data (varies by endpoint)
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        stack:
          type: string
          description: Stack trace (development only)
    ResponseMeta:
      type: object
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    CreateSourceSystem:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        config_json:
          type: object
...
security:
  - bearerAuth: []